style: update code structure
