+++ __tests__/unit/scales/band.spec.ts
@@ -1,6 +1,4 @@
-import * as d3 from 'd3-scale';
 import { Band } from '../../../src/scales/band';
-import { benchMarkBetween } from '../../test-utils/benchmark';
 import { BandOptions } from '../../../src/types';
 
 describe('band scale', () => {
@@ -136,44 +134,6 @@ describe('band scale', () => {
     const newBandScale = oldBandScale.clone();
     expect(oldBandScale.getOptions()).toStrictEqual(newBandScale.getOptions());
     expect(oldBandScale.getOptions() !== newBandScale.getOptions()).toBeTruthy();
-  });
-
-  test('compare pref with d3', async () => {
-    const domain = new Array(10000).fill('').map((item, index) => index);
-
-    const antvTest = () => {
-      const antvScale = new Band({
-        domain,
-        range: [0, 100000],
-      });
-
-      for (let i = 0; i < 100000; i += 1) {
-        antvScale.map(i);
-        if (i % 1000 === 0) {
-          antvScale.update({
-            domain: [0, 100000],
-            range: [0, 100000],
-          });
-        }
-      }
-    };
-
-    const d3Test = () => {
-      const d3Scale = d3.scaleBand().domain(domain).range([0, 100000]);
-      for (let i = 0; i < 100000; i += 1) {
-        d3Scale(i);
-        if (i % 1000 === 0) {
-          d3Scale.domain(domain).range([0, 100000]);
-        }
-      }
-    };
-    // 两者暂时不相上下，先不断言
-    await benchMarkBetween({
-      cb1: antvTest,
-      cb2: d3Test,
-      magnification: 9,
-      check: true,
-    });
   });
 
   test('test update options', () => {
