+++ __tests__/unit/scales/category.spec.ts
@@ -1,6 +1,4 @@
-import * as d3 from 'd3-scale';
 import { Category } from '../../../src/scales/category';
-import { benchMarkBetween } from '../../test-utils/benchmark';
 
 describe('category scale', () => {
   test('category has no expected defaults', () => {
@@ -118,47 +116,5 @@ describe('category scale', () => {
     expect(scale.map('苹果')).toStrictEqual('apple');
     expect(scale.invert('apple')).toStrictEqual('苹果');
     expect(scale.invert('orange')).toStrictEqual('橘子');
-  });
-
-  test('compare pref with d3', async () => {
-    // 和 d3 对大数据（十万）情况做性能对比，在调用 update 的数量较大的场景下，antv-scale 的效率较 d3 有质的提升
-    const domain = new Array(100000).fill('').map((item, index) => index);
-    const range = new Array(100000).fill('').map((item, index) => index);
-
-    const timeForAntv = () => {
-      const antvScale = new Category({
-        domain,
-        range,
-      });
-      for (let i = 0; i < 100000; i += 1) {
-        antvScale.map(i);
-        // 中途重置 domain
-        if (i % 4000 === 0) {
-          antvScale.update({
-            domain: range,
-            range,
-          });
-        }
-      }
-    };
-
-    const timeForD3 = () => {
-      const d3Scale = d3.scaleOrdinal().domain(domain).range(range);
-      for (let i = 0; i < 100000; i += 1) {
-        d3Scale(i);
-        if (i % 4000 === 0) {
-          // 中途重置 domain
-          d3Scale.domain(range).range(range);
-        }
-      }
-    };
-
-    await benchMarkBetween({
-      cb1: timeForAntv,
-      cb2: timeForD3,
-      // TODO: 这里有较大幅度的摇摆，先改为 1.0，后续优化
-      magnification: 1.0,
-      check: true,
-    });
   });
 });
