+++ __tests__/unit/tick-methods/linear.spec.ts
@@ -1,15 +1,15 @@
-import { linerTick } from '../../../src/tick-method/linear';
+import { linear } from '../../../src/tick-method/linear';
 
 describe('linear tick method', () => {
   test('test linear tick', () => {
-    const res = linerTick({
+    const res = linear({
       domain: [0, 1],
     });
     expect(res).toStrictEqual([0, 0.3, 0.5, 0.8, 1]);
   });
 
   test('test tickCount option', () => {
-    const res = linerTick({
+    const res = linear({
       domain: [0, 1],
       tickCount: 10,
     });
